requirements.txt
streamlit
llama-index
llama-index-vector-stores-chroma
chromadb
pypdf
llama-index-embeddings-openai
python-dotenv
openai
protobuf==3.20.3
